Extract solution templates manager
